requirements.txt
pymongo==4.6.1
requests>=2.31.0
selenium>=4.13.0
setuptools==69.2.0
beautifulsoup4==4.12.3
qrcode==7.4.2
Pillow==11.1.0
